accept new comments and tickets through email
Some notes:
I wasn’t able to get the email attachments going through the validator
class, so its currently doing a regex check on the extension type
rather than checking against the mime the way the uploaded attachments
do.

I need some help getting the agent auto assignment working with newly
created email tickets. I haven’t had time to get into the issue to make
it work, so currently the “default_email_agent” needs to be an agent
and not “auto”

It would probably be better if the “imap_password” setting was hashed
like a password should be but thats beyond me at the moment.